requirements.txt
alembic==1.14.0
amqp==5.3.1
APScheduler==3.11.0
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.4.0
chardet==5.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dotenv==0.9.9
Flask==3.1.0
Flask-Caching==2.3.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.5.0
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
reportlab==4.3.1
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
